removing old code and extra comments in output
